Move logsToStepMetadata to a provider component
